Add files via upload
 movies table linked with the rating table using a foreign key.
I am using Code First approach to create the database and tables. Here are the two models for the Movies and StarRatings table.
In the above model, I have captured the “Rating” total and “Rating Count” so that I can fetch them on the model binding.
The above star rating model is very simple. You can see the relationship created with the domain model Movie. I have also specified the Foreign Key to use the specific key instead of generic one.

I am using separate image files for the “filled in star” and “empty star”. The empty star gets replaced when a user hovers on it and vice versa. Here is the jQuery code to implement the hover functionality

I have the following code to generate listview in the movies controller.